+++ tests/src/class1.hpp
@@ -27,7 +27,7 @@ protected:
 class Class1 : public Base2
 {
 public:
-	Class1(int v = 5);
+	Class1(float v = 5.0f);
 	virtual ~Class1();
 	int method1(float a, float b);
 	virtual void method2(float);
